Also replaces loop over `String` with `getChars` which is hotspottable.