@anuraaga 2021-05-13
Improvement with `getChars` is more than I was expecting, ~20%. W3C propagator only changed to use `getChars`

After
```
Benchmark                                                                                Mode  Cnt     Score    Error   Units
W3CTraceContextPropagatorInjectBenchmark.measureInject                                   avgt   15    40.443 ±  0.357   ns/op
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.alloc.rate                    avgt   15  1507.718 ± 13.883  MB/sec
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.alloc.rate.norm               avgt   15    96.011 ±  0.001    B/op
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.churn.G1_Eden_Space           avgt   15  1503.231 ± 87.734  MB/sec
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.churn.G1_Eden_Space.norm      avgt   15    95.712 ±  5.268    B/op
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.churn.G1_Survivor_Space       avgt   15     0.008 ±  0.003  MB/sec
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.churn.G1_Survivor_Space.norm  avgt   15     0.001 ±  0.001    B/op
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.count                         avgt   15   153.000           counts
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.time                          avgt   15   100.000               ms
```

Before
```
Benchmark                                                                                Mode  Cnt     Score    Error   Units
W3CTraceContextPropagatorInjectBenchmark.measureInject                                   avgt   15    60.323 ±  0.657   ns/op
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.alloc.rate                    avgt   15  1010.462 ± 10.567  MB/sec
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.alloc.rate.norm               avgt   15    96.013 ±  0.001    B/op
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.churn.G1_Eden_Space           avgt   15  1009.866 ± 59.877  MB/sec
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.churn.G1_Eden_Space.norm      avgt   15    95.961 ±  5.689    B/op
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.churn.G1_Survivor_Space       avgt   15     0.005 ±  0.003  MB/sec
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.churn.G1_Survivor_Space.norm  avgt   15    ≈ 10⁻³             B/op
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.count                         avgt   15   124.000           counts
W3CTraceContextPropagatorInjectBenchmark.measureInject:·gc.time                          avgt   15    75.000               ms
```