@anuraaga 2021-05-13
Ah just in case people are curious, I'm fairly sure the `getChars` performance improvement will be mostly 0 on Java 8, the intrinsics were added in Java 9+ if I'm not mistaken, on Java 8 the loop should be the same exact bytecode.